refactor: move branches to root menu
