+++ src/views/Branches/BranchEdit.vue
@@ -22,23 +22,23 @@
             :branch="branchData"
             :branch-id="branchId"
             :assign-list="canAssign"
-            :can-edit="canEdit"
-            @isLoading="statusLoading" />
+            @isLoading="statusLoading"
+            @save="saveData"/>
       </div>
     </div>
   </main>
 </template>
 <script setup>
-import { onBeforeMount, ref } from "vue";
+import {onBeforeMount, ref} from "vue";
 import { useRouter, useRoute } from 'vue-router'
+import { notify } from "notiwind";
 import branchService from "@/services/branch.service";
 import Breadcrumbs from '@/components/Breadcrumbs.vue'
 import LoadingContent from '@/components/LoadingContent.vue'
-import BranchForm from '@/views/Parametric/Branches/BranchForm.vue'
+import BranchForm from '@/views/Branches/BranchForm.vue'
 
 const route = useRoute()
 const router = useRouter()
-const canEdit = ref(false)
 const canAssign = ref(true)
 let branchData = ref({
   "id": "",
@@ -72,23 +72,48 @@ let companyName = ref(String)
 //for breadcrumbs
 let trace = [
   { description: 'Home', pathName: 'HomePage', isLink: true, current: false },
-  { description: 'Parametric', pathName: 'ParametricPage', isLink: true, current: false },
   { description: 'Branches', pathName: 'BranchList', isLink: true, current: false },
-  { description: 'Show', pathName: 'BranchShow', isLink: false, current: true }
+  { description: 'Edit', pathName: 'BranchEdit', isLink: false, current: true }
 ]
 onBeforeMount( () => {
 
   branchId.value = route.params.branchId
 
-  branchService.get(branchId.value).then(x => {
+  branchService.edit(branchId.value).then(x => {
     branchData.value = x.data
     countryFlag.value = x.data.country.flag_url
     companyName.value = x.data.company.description
   }).catch(err => {
+
   }).finally(() => {
+
   })
 })
-
+const saveData = (event) => {
+  let data = {
+    'country': event.country,
+    'company': event.company,
+    'status': event.status
+  }
+  branchService.save(branchId.value, data).then(x => {
+    notify({
+      group: "top",
+      title: "Update",
+      text: x.data,
+      type: "success"
+    }, 5000)
+    router.push({ name: "BranchList" })
+  }).catch(err => {
+    let message = (err.code === "ERR_BAD_REQUEST") ? err.response.data.message : err.message;
+    notify({
+      group: "top",
+      title: "Error",
+      text: message,
+      type: "error"
+    }, 5000)
+  }).finally(() => {
+  })
+}
 const statusLoading = (data) => {
   loading.value = data
 }
